content-bot/requirements.txt
python-telegram-bot[ext]
google-generativeai
firebase-admin
feedparser
python-dotenv
beautifulsoup4
numpy